that was unnecessary

+++ Modules/Camouflague.cs
@@ -205,7 +205,7 @@ public static class Camouflage
     {
         target.SetColor(newOutfit.ColorId);
 
-        sender.AutoStartRpc(target.NetId, (byte)RpcCalls.SetColor)
+        sender.AutoStartRpc(target.NetId, RpcCalls.SetColor)
             .Write(target.Data.NetId)
             .Write((byte)newOutfit.ColorId)
             .EndRpc();
@@ -213,7 +213,7 @@ public static class Camouflage
         target.SetHat(newOutfit.HatId, newOutfit.ColorId);
         target.Data.DefaultOutfit.HatSequenceId += 10;
 
-        sender.AutoStartRpc(target.NetId, (byte)RpcCalls.SetHatStr)
+        sender.AutoStartRpc(target.NetId, RpcCalls.SetHatStr)
             .Write(newOutfit.HatId)
             .Write(target.GetNextRpcSequenceId(RpcCalls.SetHatStr))
             .EndRpc();
@@ -221,7 +221,7 @@ public static class Camouflage
         target.SetSkin(newOutfit.SkinId, newOutfit.ColorId);
         target.Data.DefaultOutfit.SkinSequenceId += 10;
 
-        sender.AutoStartRpc(target.NetId, (byte)RpcCalls.SetSkinStr)
+        sender.AutoStartRpc(target.NetId, RpcCalls.SetSkinStr)
             .Write(newOutfit.SkinId)
             .Write(target.GetNextRpcSequenceId(RpcCalls.SetSkinStr))
             .EndRpc();
@@ -229,7 +229,7 @@ public static class Camouflage
         target.SetVisor(newOutfit.VisorId, newOutfit.ColorId);
         target.Data.DefaultOutfit.VisorSequenceId += 10;
 
-        sender.AutoStartRpc(target.NetId, (byte)RpcCalls.SetVisorStr)
+        sender.AutoStartRpc(target.NetId, RpcCalls.SetVisorStr)
             .Write(newOutfit.VisorId)
             .Write(target.GetNextRpcSequenceId(RpcCalls.SetVisorStr))
             .EndRpc();
@@ -237,7 +237,7 @@ public static class Camouflage
         target.SetPet(newOutfit.PetId);
         target.Data.DefaultOutfit.PetSequenceId += 10;
 
-        sender.AutoStartRpc(target.NetId, (byte)RpcCalls.SetPetStr)
+        sender.AutoStartRpc(target.NetId, RpcCalls.SetPetStr)
             .Write(newOutfit.PetId)
             .Write(target.GetNextRpcSequenceId(RpcCalls.SetPetStr))
             .EndRpc();
