+++ Modules/Utils.cs
@@ -1786,7 +1786,7 @@ public static class Utils
 
                 if ((fullRpcSize <= fullRpcSizeLimit && titleRpcSizeLimit <= titleRpcSize) || title.Length <= 100)
                 {
-                    writer.AutoStartRpc(sender.NetId, (byte)RpcCalls.SetName, targetClientId)
+                    writer.AutoStartRpc(sender.NetId, RpcCalls.SetName, targetClientId)
                         .Write(sender.Data.NetId)
                         .Write(title)
                         .EndRpc();
@@ -1841,7 +1841,7 @@ public static class Utils
                     if (writer.CurrentState == CustomRpcSender.State.Finished)
                         writer = CustomRpcSender.Create("Utils.SendMessage(2)", sendOption);
 
-                    writer.AutoStartRpc(sender.NetId, (byte)RpcCalls.SetName, targetClientId)
+                    writer.AutoStartRpc(sender.NetId, RpcCalls.SetName, targetClientId)
                         .Write(sender.Data.NetId)
                         .Write(title)
                         .EndRpc();
@@ -1851,16 +1851,16 @@ public static class Utils
                         if (writer.CurrentState == CustomRpcSender.State.Finished)
                             writer = CustomRpcSender.Create("Utils.SendMessage.SendTempTitleMessage", sendOption);
 
-                        writer.AutoStartRpc(sender.NetId, (byte)RpcCalls.SetName, targetClientId)
+                        writer.AutoStartRpc(sender.NetId, RpcCalls.SetName, targetClientId)
                             .Write(sender.Data.NetId)
                             .Write(tempTitle)
                             .EndRpc();
 
-                        writer.AutoStartRpc(sender.NetId, (byte)RpcCalls.SendChat, targetClientId)
+                        writer.AutoStartRpc(sender.NetId, RpcCalls.SendChat, targetClientId)
                             .Write("\n")
                             .EndRpc();
 
-                        writer.AutoStartRpc(sender.NetId, (byte)RpcCalls.SetName, targetClientId)
+                        writer.AutoStartRpc(sender.NetId, RpcCalls.SetName, targetClientId)
                             .Write(sender.Data.NetId)
                             .Write(Main.AllPlayerNames.GetValueOrDefault(sender.PlayerId, string.Empty))
                             .EndRpc();
@@ -1920,7 +1920,7 @@ public static class Utils
                 if (shortenedText.Length > 0 && !shortenedText.IsNullOrWhiteSpace()) writer = SendMessage(shortenedText, sendTo, title, true, writer, true, sendOption: sendOption);
                 else
                 {
-                    writer.AutoStartRpc(sender.NetId, (byte)RpcCalls.SetName, targetClientId)
+                    writer.AutoStartRpc(sender.NetId, RpcCalls.SetName, targetClientId)
                         .Write(sender.Data.NetId)
                         .Write(Main.AllPlayerNames.GetValueOrDefault(sender.PlayerId, string.Empty))
                         .EndRpc();
@@ -1949,13 +1949,13 @@ public static class Utils
 
             if (writer.CurrentState == CustomRpcSender.State.Ready)
             {
-                writer.AutoStartRpc(sender.NetId, (byte)RpcCalls.SetName, targetClientId)
+                writer.AutoStartRpc(sender.NetId, RpcCalls.SetName, targetClientId)
                     .Write(sender.Data.NetId)
                     .Write(title)
                     .EndRpc();
             }
 
-            writer.AutoStartRpc(sender.NetId, (byte)RpcCalls.SendChat, targetClientId)
+            writer.AutoStartRpc(sender.NetId, RpcCalls.SendChat, targetClientId)
                 .Write(text)
                 .EndRpc();
 
@@ -1969,7 +1969,7 @@ public static class Utils
 
             if ((noSplit && final) || !noSplit)
             {
-                writer.AutoStartRpc(sender.NetId, (byte)RpcCalls.SetName, targetClientId)
+                writer.AutoStartRpc(sender.NetId, RpcCalls.SetName, targetClientId)
                     .Write(sender.Data.NetId)
                     .Write(Main.AllPlayerNames.GetValueOrDefault(sender.PlayerId, string.Empty))
                     .EndRpc();
@@ -2933,7 +2933,7 @@ public static class Utils
 
         pc.SetName(newOutfit.PlayerName);
 
-        sender.AutoStartRpc(pc.NetId, (byte)RpcCalls.SetName)
+        sender.AutoStartRpc(pc.NetId, RpcCalls.SetName)
             .Write(pc.Data.NetId)
             .Write(newOutfit.PlayerName)
             .EndRpc();
@@ -2942,7 +2942,7 @@ public static class Utils
 
         pc.SetColor(newOutfit.ColorId);
 
-        sender.AutoStartRpc(pc.NetId, (byte)RpcCalls.SetColor)
+        sender.AutoStartRpc(pc.NetId, RpcCalls.SetColor)
             .Write(pc.Data.NetId)
             .Write((byte)newOutfit.ColorId)
             .EndRpc();
@@ -2950,7 +2950,7 @@ public static class Utils
         pc.SetHat(newOutfit.HatId, newOutfit.ColorId);
         pc.Data.DefaultOutfit.HatSequenceId += 10;
 
-        sender.AutoStartRpc(pc.NetId, (byte)RpcCalls.SetHatStr)
+        sender.AutoStartRpc(pc.NetId, RpcCalls.SetHatStr)
             .Write(newOutfit.HatId)
             .Write(pc.GetNextRpcSequenceId(RpcCalls.SetHatStr))
             .EndRpc();
@@ -2958,7 +2958,7 @@ public static class Utils
         pc.SetSkin(newOutfit.SkinId, newOutfit.ColorId);
         pc.Data.DefaultOutfit.SkinSequenceId += 10;
 
-        sender.AutoStartRpc(pc.NetId, (byte)RpcCalls.SetSkinStr)
+        sender.AutoStartRpc(pc.NetId, RpcCalls.SetSkinStr)
             .Write(newOutfit.SkinId)
             .Write(pc.GetNextRpcSequenceId(RpcCalls.SetSkinStr))
             .EndRpc();
@@ -2966,7 +2966,7 @@ public static class Utils
         pc.SetVisor(newOutfit.VisorId, newOutfit.ColorId);
         pc.Data.DefaultOutfit.VisorSequenceId += 10;
 
-        sender.AutoStartRpc(pc.NetId, (byte)RpcCalls.SetVisorStr)
+        sender.AutoStartRpc(pc.NetId, RpcCalls.SetVisorStr)
             .Write(newOutfit.VisorId)
             .Write(pc.GetNextRpcSequenceId(RpcCalls.SetVisorStr))
             .EndRpc();
@@ -2974,7 +2974,7 @@ public static class Utils
         pc.SetPet(newOutfit.PetId);
         pc.Data.DefaultOutfit.PetSequenceId += 10;
 
-        sender.AutoStartRpc(pc.NetId, (byte)RpcCalls.SetPetStr)
+        sender.AutoStartRpc(pc.NetId, RpcCalls.SetPetStr)
             .Write(newOutfit.PetId)
             .Write(pc.GetNextRpcSequenceId(RpcCalls.SetPetStr))
             .EndRpc();
@@ -2982,7 +2982,7 @@ public static class Utils
         pc.SetNamePlate(newOutfit.NamePlateId);
         pc.Data.DefaultOutfit.NamePlateSequenceId += 10;
 
-        sender.AutoStartRpc(pc.NetId, (byte)RpcCalls.SetNamePlateStr)
+        sender.AutoStartRpc(pc.NetId, RpcCalls.SetNamePlateStr)
             .Write(newOutfit.NamePlateId)
             .Write(pc.GetNextRpcSequenceId(RpcCalls.SetNamePlateStr))
             .EndRpc();
