+++ Roles/Neutral/Spiritcaller.cs
@@ -119,16 +119,16 @@ public class Spiritcaller : RoleBase
             var writer = CustomRpcSender.Create("SpiritCallerSendMessage", SendOption.Reliable);
             writer.StartMessage(target.OwnerId);
 
-            writer.StartRpc(target.NetId, (byte)RpcCalls.SetName)
+            writer.StartRpc(target.NetId, RpcCalls.SetName)
                 .Write(target.Data.NetId)
                 .Write(GetString("SpiritcallerNoticeTitle"))
                 .EndRpc();
 
-            writer.StartRpc(target.NetId, (byte)RpcCalls.SendChat)
+            writer.StartRpc(target.NetId, RpcCalls.SendChat)
                 .Write(GetString("SpiritcallerNoticeMessage"))
                 .EndRpc();
 
-            writer.StartRpc(target.NetId, (byte)RpcCalls.SetName)
+            writer.StartRpc(target.NetId, RpcCalls.SetName)
                 .Write(target.Data.NetId)
                 .Write(target.Data.PlayerName)
                 .EndRpc();
